# web/lib/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AapatMitra API",
  "version": "1.0.0"
 },
 "paths": {
  "/api/v1/auth/otp/request": {
   "post": {
    "tags": [
     "identity"
    ],
    "summary": "Otp Request",
    "operationId": "otp_request_api_v1_auth_otp_request_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpRequestIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OtpRequestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/otp/verify": {
   "post": {
    "tags": [
     "identity"
    ],
    "summary": "Otp Verify",
    "operationId": "otp_verify_api_v1_auth_otp_verify_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OtpVerifyIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/refresh": {
   "post": {
    "tags": [
     "identity"
    ],
    "summary": "Refresh",
    "operationId": "refresh_api_v1_auth_refresh_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RefreshIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "tags": [
     "identity"
    ],
    "summary": "Logout",
    "operationId": "logout_api_v1_auth_logout_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/LogoutIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/me": {
   "get": {
    "tags": [
     "identity"
    ],
    "summary": "Me",
    "operationId": "me_api_v1_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "identity"
    ],
    "summary": "Patch Me",
    "operationId": "patch_me_api_v1_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MePatchIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/me/devices/{device_id}": {
   "delete": {
    "tags": [
     "identity"
    ],
    "summary": "Revoke My Device",
    "operationId": "revoke_my_device_api_v1_me_devices__device_id__delete",
    "parameters": [
     {
      "name": "device_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Device Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/me/devices/{device_id}/push-token": {
   "put": {
    "tags": [
     "identity"
    ],
    "summary": "Push Token",
    "operationId": "push_token_api_v1_me_devices__device_id__push_token_put",
    "parameters": [
     {
      "name": "device_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Device Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PushTokenIn"
       }
      }
     }
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/me/leaderboard-opt-out": {
   "post": {
    "tags": [
     "identity"
    ],
    "summary": "Leaderboard Opt Out",
    "operationId": "leaderboard_opt_out_api_v1_me_leaderboard_opt_out_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OptOutIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "DeviceIn": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "platform": {
      "type": "string",
      "enum": [
       "android",
       "web"
      ],
      "title": "Platform",
      "default": "android"
     },
     "appVersion": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 40
       },
       {
        "type": "null"
       }
      ],
      "title": "Appversion"
     },
     "osVersion": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 40
       },
       {
        "type": "null"
       }
      ],
      "title": "Osversion"
     },
     "model": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 80
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "fcmToken": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 4096
       },
       {
        "type": "null"
       }
      ],
      "title": "Fcmtoken"
     },
     "publicKeyEd25519": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 400
       },
       {
        "type": "null"
       }
      ],
      "title": "Publickeyed25519"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "id"
    ],
    "title": "DeviceIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LogoutIn": {
    "properties": {
     "allDevices": {
      "type": "boolean",
      "title": "Alldevices",
      "default": false
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "LogoutIn"
   },
...
   "MePatchIn": {
    "properties": {
     "preferredLanguage": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[a-z]{2,3}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Preferredlanguage"
     },
     "textScale": {
      "anyOf": [
       {
        "type": "integer",
        "enum": [
         100,
         130,
         160
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Textscale"
     },
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 80,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "MePatchIn"
   },
...
   "OptOutIn": {
    "properties": {
     "optOut": {
      "type": "boolean",
      "title": "Optout"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "optOut"
    ],
    "title": "OptOutIn"
   },
   "OtpRequestIn": {
    "properties": {
     "phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "staffId": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 40
       },
       {
        "type": "null"
       }
      ],
      "title": "Staffid"
     },
     "purpose": {
      "type": "string",
      "enum": [
       "login",
       "register",
       "recovery"
      ],
      "title": "Purpose",
      "default": "login"
     },
     "appHash": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 20
       },
       {
        "type": "null"
       }
      ],
      "title": "Apphash"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "OtpRequestIn"
   },
   "OtpRequestOut": {
    "properties": {
     "sent": {
      "type": "boolean",
      "title": "Sent"
     },
     "challengeId": {
      "type": "string",
      "format": "uuid",
      "title": "Challengeid"
     },
     "expiresAt": {
      "type": "string",
      "title": "Expiresat"
     },
     "resendAfterS": {
      "type": "integer",
      "title": "Resendafters"
     }
    },
    "type": "object",
    "required": [
     "sent",
     "challengeId",
     "expiresAt",
     "resendAfterS"
    ],
    "title": "OtpRequestOut"
   },
   "OtpVerifyIn": {
    "properties": {
     "challengeId": {
      "type": "string",
      "format": "uuid",
      "title": "Challengeid"
     },
     "otp": {
      "type": "string",
      "pattern": "^\\d{6}$",
      "title": "Otp"
     },
     "device": {
      "$ref": "#/components/schemas/DeviceIn"
     },
     "role": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "patient",
         "asha",
         "volunteer",
         "doctor",
         "facility_staff",
         "district_admin"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     },
     "registration": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Registration"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "challengeId",
     "otp",
     "device"
    ],
    "title": "OtpVerifyIn"
   },
...
   "PushTokenIn": {
    "properties": {
     "fcmToken": {
      "type": "string",
      "maxLength": 4096,
      "title": "Fcmtoken"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "fcmToken"
    ],
    "title": "PushTokenIn"
   },
...
   "RefreshIn": {
    "properties": {
     "refreshToken": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Refreshtoken"
     },
     "deviceId": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deviceid"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "RefreshIn"
   },
...
   "TokenOut": {
    "properties": {
     "accessToken": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accesstoken"
     },
     "accessTokenExpiresAt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accesstokenexpiresat"
     },
     "refreshToken": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Refreshtoken"
     },
     "refreshTokenExpiresAt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Refreshtokenexpiresat"
     },
     "user": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/UserOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "mfaRequired": {
      "type": "boolean",
      "title": "Mfarequired",
      "default": false
     },
     "chooseRole": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Chooserole"
     },
     "serverTime": {
      "type": "string",
      "title": "Servertime"
     }
    },
    "type": "object",
    "required": [
     "serverTime"
    ],
    "title": "TokenOut"
   },
...
   "UserOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "role": {
      "type": "string",
      "enum": [
       "patient",
       "asha",
       "volunteer",
       "doctor",
       "facility_staff",
       "district_admin"
      ],
      "title": "Role"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "phoneMasked": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phonemasked"
     },
     "staffId": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Staffid"
     },
     "preferredLanguage": {
      "type": "string",
      "title": "Preferredlanguage"
     },
     "textScale": {
      "type": "integer",
      "title": "Textscale"
     },
     "districtCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Districtcode"
     },
     "homeVillageId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Homevillageid"
     },
     "householdId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Householdid"
     },
     "patientId": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Patientid"
     },
     "facilityIds": {
      "anyOf": [
       {
        "items": {
         "type": "string",
         "format": "uuid"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Facilityids"
     },
     "villageIds": {
      "anyOf": [
       {
        "items": {
         "type": "string",
         "format": "uuid"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Villageids"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     }
    },
    "type": "object",
    "required": [
     "id",
     "role",
     "status",
     "preferredLanguage",
     "textScale",
     "version"
    ],
    "title": "UserOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}